+++ hxcpp-debugger-protocol/src/fallback/java/debugger/ExceptionInfo.java
@@ -1,0 +1,5 @@
+package debugger;
+
+public class ExceptionInfo {
+  public String text;
+}